Refactored the analytics out of the homescreen into separate files
UI Refactor
