+++ feature/home/src/main/java/com/scrolless/app/feature/home/UsageAnalyticsModels.kt
@@ -1,0 +1,47 @@
+/*
+ * Copyright (C) 2026 Scrolless
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package com.scrolless.app.feature.home
+
+import androidx.compose.runtime.Immutable
+import com.scrolless.app.core.model.BlockableApp
+import com.scrolless.app.core.model.SessionSegment
+import com.scrolless.app.core.model.usage.WeekdayUsageAverage
+import java.time.LocalDate
+import java.time.ZonedDateTime
+
+@Immutable
+data class UsageAnalyticsUiState(
+    val selectedDate: LocalDate = ZonedDateTime.now().toLocalDate(),
+    val today: LocalDate = ZonedDateTime.now().toLocalDate(),
+    val dailyTotalMillis: Long = 0L,
+    val sessionSegments: List<SessionSegment> = emptyList(),
+    val appTotals: List<AppUsageTotal> = emptyList(),
+    val daySummaries: Map<LocalDate, UsageAnalyticsDayUiState> = emptyMap(),
+    val weekdayAverages: List<WeekdayUsageAverage> = emptyList(),
+    val canNavigateNext: Boolean = false,
+)
+
+@Immutable
+data class UsageAnalyticsDayUiState(
+    val date: LocalDate,
+    val dailyTotalMillis: Long = 0L,
+    val sessionSegments: List<SessionSegment> = emptyList(),
+    val appTotals: List<AppUsageTotal> = emptyList(),
+)
+
+@Immutable
+data class AppUsageTotal(val app: BlockableApp, val totalMillis: Long)
